139 Extract object functions to own classes
