+++ src/logger/CreateTimestamp.ts
@@ -1,0 +1,3 @@
+export function createTimestamp(): string {
+    return new Date().toISOString()
+}
