+++ src/request/GetContentType.ts
