+++ tests/request/ExtractInformationFromRequest.test.ts
@@ -5,9 +5,11 @@ import {
     usersRequestWithoutVariables
 } from '../ExampleSchemas'
 import {generateGetParametersFromGraphQLRequestInfo} from '../TestHelpers'
-import {DefaultRequestInformationExtractor} from '~/src'
-
-const requestInformationExtractor = new DefaultRequestInformationExtractor()
+import {
+    extractInformationFromBody,
+    extractInformationFromRequest,
+    extractInformationFromUrlParameters
+} from '~/src'
 
 describe('Test that request information is extracted correctly from url parameters', () => {
     test.each`
@@ -17,7 +19,7 @@ describe('Test that request information is extracted correctly from url paramete
     ${usersRequestWithoutVariables}      | ${usersRequestWithoutVariables.query}      | ${usersRequestWithoutVariables.variables}                      | ${usersRequestWithoutVariables.operationName}  
     `('expects for request $request to extract values correctly', ({request, expectedQuery, expectedVariables, expectedOperationName}) => {
         const requestUrl = `http://doesnotmatter.com/graphql?${generateGetParametersFromGraphQLRequestInfo(request)}`
-        const result = requestInformationExtractor.extractInformationFromUrlParameters(requestUrl)
+        const result = extractInformationFromUrlParameters(requestUrl)
         expect(result.query).toBe(expectedQuery)
         expect(result.variables).toBe(expectedVariables)
         expect(result.operationName).toBe(expectedOperationName)
@@ -31,7 +33,7 @@ test('Get fitting error if body type contains invalid type', () => {
         body: true,
         method: 'POST'
     }
-    const response = requestInformationExtractor.extractInformationFromBody(request)
+    const response = extractInformationFromBody(request)
     expect(response.error?.graphQLError.message).toBe('POST body contains invalid type boolean. Only "object" and "string" are supported.')
 })
 
@@ -41,7 +43,7 @@ test('Get fitting error if body contains a Buffer', () => {
         url: 'doesnotmatter',
         body: Buffer.alloc(3) ,
     }
-    const response = requestInformationExtractor.extractInformationFromBody(request)
+    const response = extractInformationFromBody(request)
     expect(response.error?.graphQLError.message).toBe('Cannot extract information from body because it contains an object buffer!')
 })
 
@@ -54,7 +56,7 @@ test('Should properly extract variables from url', () => {
         url: '/graphql?query=mutation&variables=findme',
         body: { query: 'doesnotmatter'} ,
     }
-    const response = requestInformationExtractor.extractInformationFromRequest(request)
+    const response = extractInformationFromRequest(request)
     expect(response.variables).toBe('findme')
 })
 
@@ -67,7 +69,7 @@ test('Should properly extract query from body for graphql request', () => {
         url: 'pengpeng',
         body: { query: 'findTheQuery'}
     }
-    const response = requestInformationExtractor.extractInformationFromRequest(request)
+    const response = extractInformationFromRequest(request)
     expect(response.query).toBe('{"query":"findTheQuery"}')
 })
 
@@ -79,6 +81,6 @@ test('Should read body even if url is not set', () => {
         },
         body: { query: 'findTheQuery'},
     }
-    const response = requestInformationExtractor.extractInformationFromRequest(request)
+    const response = extractInformationFromRequest(request)
     expect(response.query).toBe('{"query":"findTheQuery"}')
 })
